+++ crates/edr_eth/src/transaction/type.rs
@@ -1,3 +1,5 @@
+use super::request::Eip7702;
+
 /// The type of transaction.
 #[repr(u64)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, PartialOrd, Ord)]
@@ -6,6 +8,7 @@ pub enum TransactionType {
     Eip2930 = 1,
     Eip1559 = 2,
     Eip4844 = 3,
+    Eip7702 = Eip7702::TYPE as u64,
 }
 
 impl From<TransactionType> for u64 {