+++ crates/edr_test_utils/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod env;
+pub mod secret_key;
 mod fd_lock;
 
 pub use fd_lock::{new_fd_lock, RwLock};