+++ crates/foundry/evm/core/src/contracts.rs
@@ -37,6 +37,8 @@ impl From<foundry_compilers::ArtifactId> for ArtifactId {
             name,
             source,
             version,
+            build_id: _,
+            profile: _,
         } = value;
 
         Self {